This was missing from merge for some reason (#10)
(cherry picked from commit 305314d)
