+++ config/locales/en.yml
@@ -55,6 +55,12 @@ en:
           detail: "As you have stated you have a combined monthly income %{total_income}, you would need to contribute %{contribution} of your stated %{fee} fee or would receive a partial refund of %{remission} if you have paid the full fee within the last 3 months"
         sharing_income:
           detail: "As you and your partner have stated you have a combined monthly income %{total_income}, you would need to contribute %{contribution} of your stated %{fee} fee or would receive a partial refund of %{remission} if you have paid the full fee within the last 3 months"
+      no_remission:
+        header: "You're unlikely to get help with your fee"
+        single:
+          detail: "With a fee of %{fee} and combined monthly income of %{total_income}, it is unlikely that you'll be able to get financial help, unless you are likely to experience exceptional hardship (see guide)"
+        sharing_income:
+          detail: "With a fee of %{fee} and combined monthly income of %{total_income}, it is unlikely that you'll be able to get financial help, unless you are likely to experience exceptional hardship (see guide)"
       explanation_suffix: "as long as you"
       explanation_suffix_joining_word: "or"
       explanation_suffix_fields:
