+++ test_common/messaging/en.yml
@@ -33,6 +33,15 @@ en:
         positive:
           heading: You should be eligible for a Partial Remission
           detail: As you and your partner have stated you have a combined monthly income %{total_income}, you would need to contribute %{contribution} of your stated %{fee} fee or would receive a partial refund of %{remission} if you have paid the full fee within the last 3 months
+    no_remission:
+      single:
+        heading: You're unlikely to get help with your fee
+        negative:
+          detail: With a fee of %{fee} and combined monthly income of %{total_income}, it is unlikely that you'll be able to get financial help, unless you are likely to experience exceptional hardship (see guide)
+      married:
+        heading: You're unlikely to get help with your fee
+        negative:
+          detail: With a fee of %{fee} and combined monthly income of %{total_income}, it is unlikely that you'll be able to get financial help, unless you are likely to experience exceptional hardship (see guide)
   hwf_pages:
     total_income:
       errors:
